Migrate project to TypeScript, Vite, and update dependencies
