+++ package.json
@@ -3,6 +3,7 @@
   "homepage": "nextsuite.wach.id",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.4",
     "@fullcalendar/bootstrap5": "^5.11.3",
@@ -48,36 +49,29 @@
     "reactstrap": "^9.1.5",
     "simplebar-react": "^2.4.3",
     "slick-carousel": "^1.8.1",
-    "source-map-explorer": "^2.5.2",
     "styled-components": "^5.3.6",
     "sweetalert2": "^11.4.8"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "react-scripts": "^5.0.1",
-    "sass": "1.56.1",
-    "web-vitals": "^4.2.4",
-    "workbox-core": "^7.3.0",
-    "workbox-expiration": "^7.3.0",
-    "workbox-precaching": "^7.3.0",
-    "workbox-routing": "^7.3.0",
-    "workbox-strategies": "^7.3.0"
+    "@types/node": "^22.14.3",
+    "@types/react": "^18.3.18",
+    "@types/react-beautiful-dnd": "^13.1.8",
+    "@types/react-dom": "^18.3.5",
+    "@types/react-helmet": "^6.1.11",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-select": "^5.0.1",
+    "@types/react-slick": "^0.23.13",
+    "@vitejs/plugin-react": "^4.7.0",
+    "sass": "^1.89.2",
+    "typescript": "^4.9.5",
+    "vite": "^7.0.6",
+    "web-vitals": "^4.2.4"
   },
   "scripts": {
-    "analyze": "source-map-explorer 'build/static/js/*.js'",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "type-check": "tsc --noEmit"
   },
   "browserslist": {
     "production": [
