+++ src/App.tsx
@@ -1,12 +1,13 @@
+import React from "react";
 import Router from "./route/Index";
-
 import ThemeProvider from "./layout/provider/Theme";
 
-const App = () => {
+const App: React.FC = () => {
   return (
     <ThemeProvider>
       <Router />
     </ThemeProvider>
   );
 };
+
 export default App;
