+++ src/main.tsx
@@ -9,13 +9,16 @@ import "./assets/scss/style-email.scss";
 
 import reportWebVitals from "./reportWebVitals";
 
-const root = createRoot(document.getElementById('root'));
+const container = document.getElementById('root');
+if (!container) throw new Error('Failed to find the root element');
+
+const root = createRoot(container);
 root.render(
-  <>
+  <React.StrictMode>
     <BrowserRouter>
       <App />
     </BrowserRouter>
-  </>
+  </React.StrictMode>
 );
 
 // If you want your app to work offline and load faster, you can change
